fix: changing product price to decimal

+++ back-end/src/database/migrations/20220808183050-create-products.js
@@ -15,7 +15,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(4,2)
       },
       url_image: {
         allowNull: false,
